refactor(kornia-image): merge color_space (ColorSpace + DynImage) into color_spaces; single color module
Drops the confusingly-named singular color_space.rs; ColorSpace, DynImage and the
typed newtypes now all live in color_spaces.rs. Top-level re-exports unchanged
(kornia_image::ColorSpace / DynImage still resolve).
